# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartSchool Asset API",
        "description": "Dokumentasi API untuk sistem manajemen aset SmartSchool. Mencakup peminjaman barang, pengambilan konsumsi, barang masuk, dan laporan.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/item-usages": {
            "get": {
                "tags": [
                    "Pengambilan Konsumsi"
                ],
                "summary": "Riwayat pengambilan konsumsi milik sendiri",
                "description": "Menampilkan daftar pengambilan barang konsumsi yang dilakukan oleh user yang sedang login.",
                "operationId": "f8f5d076c91ea1372785e431d2ffd1ec",
                "responses": {
                    "200": {
                        "description": "Halaman riwayat pengambilan"
                    },
                    "401": {
                        "description": "Belum login"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pengambilan Konsumsi"
                ],
                "summary": "Simpan pengambilan konsumsi (user login)",
                "description": "Menyimpan pengambilan barang konsumsi oleh user yang sedang login dan otomatis mengurangi stok.",
                "operationId": "20959bc5c5dc00902a4d15fa4dc47438",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_barang",
                                    "jumlah_ambil"
                                ],
                                "properties": {
                                    "id_barang": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "jumlah_ambil": {
                                        "description": "Jumlah yang diambil (min: 1)",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect ke riwayat pengambilan jika berhasil"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/item-usages": {
            "get": {
                "tags": [
                    "Pengambilan Konsumsi"
                ],
                "summary": "Daftar semua pengambilan konsumsi (admin)",
                "description": "Menampilkan semua pengambilan konsumsi dengan filter tanggal, nama pengambil, dan jenis barang. Hanya untuk admin.",
                "operationId": "3b79af0719e1ad64e6ba1feefc212b6e",
                "parameters": [
                    {
                        "name": "dari",
                        "in": "query",
                        "description": "Tanggal mulai (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sampai",
                        "in": "query",
                        "description": "Tanggal akhir (Y-m-d)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "nama",
                        "in": "query",
                        "description": "Nama pengambil",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "barang",
                        "in": "query",
                        "description": "ID barang",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Halaman daftar pengambilan admin"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/item-usages/export": {
            "get": {
                "tags": [
                    "Pengambilan Konsumsi"
                ],
                "summary": "Export pengambilan konsumsi ke Excel",
                "description": "Download file Excel berisi data pengambilan konsumsi. Mendukung filter yang sama dengan halaman daftar. Hanya untuk admin.",
                "operationId": "cac412e0f7bd68e791af8c26a1d9f1dd",
                "parameters": [
                    {
                        "name": "dari",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sampai",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "nama",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "barang",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File Excel (.xlsx)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/item-usages/{id}": {
            "get": {
                "tags": [
                    "Pengambilan Konsumsi"
                ],
                "summary": "Form pengambilan barang konsumsi (user login)",
                "description": "Menampilkan form pengambilan barang konsumsi. Barang aset tidak bisa diambil lewat sini.",
                "operationId": "f60fd2280ee4822f0a0a5ca3b00a626f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID barang konsumsi",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Halaman form pengambilan"
                    },
                    "302": {
                        "description": "Redirect dengan error jika barang bukan konsumsi"
                    },
                    "404": {
                        "description": "Barang tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/ambil": {
            "get": {
                "tags": [
                    "Pengambilan Konsumsi"
                ],
                "summary": "Form pengambilan konsumsi publik (tanpa login)",
                "description": "Menampilkan form pengambilan barang konsumsi yang dapat diakses tanpa login. Dipanggil setelah scan QR barang.",
                "operationId": "e305700a866ac8c3f556c51a6f8b1f8c",
                "responses": {
                    "200": {
                        "description": "Halaman form pengambilan publik"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pengambilan Konsumsi"
                ],
                "summary": "Proses pengambilan konsumsi publik (tanpa login)",
                "description": "Memproses pengambilan barang konsumsi tanpa memerlukan login. Bisa mengambil lebih dari satu barang sekaligus dalam satu transaksi.",
                "operationId": "ea70a87358d7b0b62be6af6c23693801",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_pengambil",
                                    "sebagai",
                                    "items"
                                ],
                                "properties": {
                                    "nama_pengambil": {
                                        "type": "string",
                                        "example": "Ahmad Fauzi"
                                    },
                                    "sebagai": {
                                        "type": "string",
                                        "enum": [
                                            "murid",
                                            "pegawai"
                                        ],
                                        "example": "murid"
                                    },
                                    "items": {
                                        "description": "Minimal 1 barang",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "id_barang": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "jumlah_ambil": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect ke halaman sukses jika berhasil"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token autentikasi Laravel Sanctum/Session",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}